+++ src/CoreEx.EntityFrameworkCore/EfDbArgs.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Avanade. Licensed under the MIT License. See https://github.com/Avanade/CoreEx
 
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.ChangeTracking;
 
 namespace CoreEx.EntityFrameworkCore
 {
@@ -25,7 +26,7 @@ namespace CoreEx.EntityFrameworkCore
         }
 
         /// <summary>
-        /// Indicates that the underlying <see cref="DbContext"/> <see cref="DbContext.SaveChanges()"/> is to be performed automatically (defauls to <c>true</c>);
+        /// Indicates that the underlying <see cref="DbContext"/> <see cref="DbContext.SaveChanges()"/> is to be performed automatically. Defaults to <c>true</c>.
         /// </summary>
         public bool SaveChanges { get; set; } = true;
 
@@ -33,5 +34,17 @@ namespace CoreEx.EntityFrameworkCore
         /// Indicates whether the data should be refreshed (reselected where applicable) after a <b>save</b> operation (defaults to <c>true</c>); is dependent on <see cref="SaveChanges"/> being performed.
         /// </summary>
         public bool Refresh { get; set; } = true;
+
+        /// <summary>
+        /// Indicates whether the <see cref="EfDbQuery{T, TModel}"/> will <i>not</i> track entities (see <see cref="EntityFrameworkQueryableExtensions.AsNoTracking{TEntity}(IQueryable{TEntity})"/>. Defaults to <c>true</c> in that the queried entities will not be tracked.
+        /// </summary>
+        public bool QueryNoTracking { get; set; } = true;
+
+        /// <summary>
+        /// Indicates whether the <see cref="IEfDb.GetAsync{T, TModel}(EfDbArgs, Entities.CompositeKey, CancellationToken)"/> performs a <see cref="ChangeTracker.Clear"/> such that the retrieved entity is not tracked. Defaults to <c>false</c>.
+        /// </summary>
+        /// <remarks>The <see cref="EfDb{TDbContext}.GetAsync{T, TModel}(EfDbArgs, Entities.CompositeKey, CancellationToken)"/> implementation performs a <see cref="Microsoft.EntityFrameworkCore.DbContext.FindAsync{TEntity}(object?[], CancellationToken)"/> 
+        /// internally which automatically attaches and tracks the retrieved entity.</remarks>
+        public bool ClearChangeTrackerAfterGet { get; set; } = false;
     }
 }