+++ src/CoreEx/Http/Extended/TypedMappedHttpClient.cs
@@ -17,8 +17,6 @@ namespace CoreEx.Http.Extended
     /// </summary>
     public sealed class TypedMappedHttpClient : TypedMappedHttpClientCore<TypedMappedHttpClient>
     {
-        private readonly Func<HttpRequestMessage, CancellationToken, Task>? _onBeforeRequest;
-
         /// <summary>
         /// Initializes a new instance of the <see cref="TypedHttpClientCore{TBase}"/>.
         /// </summary>
@@ -36,15 +34,6 @@ namespace CoreEx.Http.Extended
             executionContext ?? (ExecutionContext.HasCurrent ? ExecutionContext.Current : new ExecutionContext()),
             settings ?? ExecutionContext.GetService<SettingsBase>() ?? new DefaultSettings(),
             logger ?? ExecutionContext.GetService<ILogger<TypedMappedHttpClient>>() ?? NullLoggerFactory.Instance.CreateLogger<TypedMappedHttpClient>()) 
-            => _onBeforeRequest = onBeforeRequest;
-
-        /// <inheritdoc/>
-        protected async override Task OnBeforeRequest(HttpRequestMessage request, CancellationToken cancellationToken)
-        {
-            if (_onBeforeRequest != null)
-                await _onBeforeRequest(request, cancellationToken).ConfigureAwait(false);
-
-            await base.OnBeforeRequest(request, cancellationToken).ConfigureAwait(false);
-        }
+            => OnBeforeRequest(onBeforeRequest);
     }
 }